// packages/@essence/essence-constructor-classes/src/FieldPopover/stores/FieldPopoverModel/FieldPopoverModel.types.ts
import {IStoreBaseModel, IRecordsModel, IRecord} from "@essence-community/constructor-share/types";

export interface IFieldPopoverModel extends IStoreBaseModel {
    valueFields: Array<[string, string]>;
    valueField: string;
    recordsStore: IRecordsModel;
    getLabel: (rec: IRecord) => string;
}